Migrate to sbt
