--- .mill-version
@@ -1,1 +1,0 @@
-0.12.9
