+++ .sbtopts
@@ -1,0 +1,1 @@
+-J-Xmx6g
