+++ modules/core/src/main/scala/jsonrpclib/Channel.scala
