+++ modules/core/src/main/scala/jsonrpclib/ConflictingMethodError.scala
