+++ modules/core/src/main/scala/jsonrpclib/Endpoint.scala
