+++ modules/core/src/main/scala/jsonrpclib/ErrorCodec.scala
