+++ modules/core/src/main/scala/jsonrpclib/ErrorPayload.scala
