+++ modules/core/src/main/scala/jsonrpclib/ErrorReport.scala
