+++ modules/core/src/main/scala/jsonrpclib/Message.scala
