+++ modules/core/src/main/scala/jsonrpclib/Monadic.scala
