+++ modules/core/src/main/scala/jsonrpclib/ProtocolError.scala
