+++ modules/core/src/main/scala/jsonrpclib/StubTemplate.scala
