+++ modules/core/src/main/scala/jsonrpclib/internals/Constants.scala
