+++ modules/core/src/main/scala/jsonrpclib/internals/FutureBaseChannel.scala
