+++ modules/core/src/main/scala/jsonrpclib/internals/RawMessage.scala
