+++ modules/fs2/src/main/scala/jsonrpclib/fs2/CancelTemplate.scala
